requirements.txt
requests
pydub
python-dotenv